Condition work properly for the first event

+++ SquibblesMania/Assets/Scripts/Event/ConditionReleaseEvent.cs
@@ -31,7 +31,7 @@ public class Condition
    public ConditionType conditionType;
    public ColorToChose colorsToChose;
  
-   public int numberOfSteps;
+   public float numberOfSteps;
    
    public string conditionToRelease;
 }
